checkmarx/requirements.txt
fastapi==0.63.0
gunicorn==20.1.0
numpy~=1.20.2
opencv-python-headless==4.5.1.48
Pillow==8.2.0
pydantic==1.8.1
python-multipart==0.0.5
starlette==0.13.6
uvicorn==0.13.4
